+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.1.0 (Next)
-- Update `instances` var to `map` to support Ansible group specification.
+- Update `instances` var to `map(set(object))` to support Ansible group specification.
+- Support Ansible inventory groups for `instances` variable.
 
 ### 1.0.2
 - Update Azure instances to use `name` argument from resource.